refactor(widget): split DataGrid into modules
Split datagrid.rs (4754 lines) into submodules:
- mod.rs: DataGrid struct and implementations (2252 lines)
- types.rs: standalone types (535 lines)
- tests.rs: all tests (2003 lines)
